+++ drivers/dfu/dfu.c
@@ -357,7 +357,7 @@ int dfu_config_entities(char *env, char *interface, int num)
 	dfu_alt_num = dfu_find_alt_num(env);
 	debug("%s: dfu_alt_num=%d\n", __func__, dfu_alt_num);
 
-	dfu = calloc(sizeof(*dfu), dfu_alt_num);
+	dfu = calloc(sizeof(*dfu), (size_t)dfu_alt_num); /* SECURITY FIX: Cast to size_t */
 	if (!dfu)
 		return -1;
 	for (i = 0; i < dfu_alt_num; i++) {
